+++ src/types.ts
@@ -1,0 +1,7 @@
+import type { RuleConfig as PreferredExportDefaultNamingRuleConfig } from "./preferred-export-default-naming";
+import type { RuleConfig as PreventImportsRuleConfig } from "./prevent-imports";
+
+export type Configs = {
+  "preferred-export-default-naming": PreferredExportDefaultNamingRuleConfig;
+  "prevent-imports": PreventImportsRuleConfig;
+};